+++ packages/namespaces/src/index.js
@@ -4,5 +4,7 @@ export { default as getNamespaceByAction } from './getNamespaceByAction';
 export { default as isActionFromNamespace } from './isActionFromNamespace';
 export * from './constants';
 export { default as getStateByAction } from './getStateByAction';
+export { default as getStateByFeatureAndAction } from './getStateByFeatureAndAction';
 export { default as getStateByNamespace } from './getStateByNamespace';
+export { default as getStateByFeatureAndNamespace } from './getStateByFeatureAndNamespace';
 export { default as preventNamespace } from './preventNamespace';